rename: change filename 'constant.ts' to 'image.ts'
